feat: ErrorLogger 구현
