+++ 14th-team5-iOS/Core/Sources/Bibbi/BBErrorLogger.swift
@@ -20,21 +20,3 @@ extension BBErrorLogger {
     public func log<E>(localizedError error: E) where E: LocalizedError { }
     public func log(data: Data, response: URLResponse) { }
 }
-
-
-// MARK: - APIWorker Error Logger
-
-public struct APIWorkerErrorLogger: BBErrorLogger {
-    
-    public init() { }
-    
-    /// 매개변수로 주어진 `Error`의 로그를 출력합니다.
-    /// - Parameter error: `Error` 프로토콜을 준수하는 에러입니다.
-    public func log<E>(localizedError error: E) where E: LocalizedError {
-        var errorLog: String = "-- [APIWorker Error Log] ----------------------------------\n"
-        let description = " - DESCRIPTION: \(error.localizedDescription)\n"
-        errorLog.append(description)
-        print(errorLog + "----------------------------------------------------------\n")
-    }
-    
-}
