+++ 14th-team5-iOS/Core/Sources/Bibbi/BBNetwork/Network/BBNetworkErrorLogger.swift
@@ -1,0 +1,16 @@
+//
+//  BBNetworkErrorLogger.swift
+//  Core
+//
+//  Created by 김건우 on 10/9/24.
+//
+
+import Foundation
+
+public struct BBNetworkErrorLogger: BBErrorLogger {
+    
+    public init() { }
+    
+    public func log<E>(localizedError error: E) where E : LocalizedError { }
+    
+}
